requirements.txt
# Instagram Reels Analysis - Requirements

# Core dependencies
instagrapi>=1.18.0
gradio>=4.0.0
transformers>=4.30.0
torch>=2.0.0
torchvision
torchaudio

# Data processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
datasets>=2.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Text processing
emoji>=2.0.0
regex

# Utilities
python-dotenv
google-colab  # For Colab secrets (if running in Colab)

# Optional: For better performance
# accelerate>=0.20.0
# bitsandbytes>=0.40.0
